refactor: upgrade to LibLog 4.2 and privatise NullLogProvider
