+++ src/ScriptCs.Contracts/ScriptCs.Contracts.csproj
@@ -45,7 +45,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="App_Packages\LibLog.4.1\LibLog.cs" />
+    <Compile Include="App_Packages\LibLog.4.2\LibLog.cs" />
     <Compile Include="AssemblyReferences.cs" />
     <Compile Include="BehaviorAfterCode.cs" />
     <Compile Include="..\ScriptCs.Core\Guard.cs">
@@ -105,7 +105,6 @@
     <Compile Include="LogLevel.cs" />
     <Compile Include="LogProviderExtensions.cs" />
     <Compile Include="ModuleAttribute.cs" />
-    <Compile Include="NullLogProvider.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ScriptPack.cs" />
     <Compile Include="ScriptPackSession.cs" />
