+++ src/ScriptCs.Contracts/packages.config
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="LibLog" version="4.1.1" targetFramework="net45" developmentDependency="true" />
+  <package id="LibLog" version="4.2.0" targetFramework="net45" developmentDependency="true" />
   <package id="StyleCop.MSBuild" version="4.7.49.0" targetFramework="net45" developmentDependency="true" />
 </packages>
